[Agent Builder] Colocate Raw Vega integrity checks with chart types
