+++ x-pack/platform/packages/shared/agent-builder/agent-builder-visualizations-server/vega/chart_types/integrity_utils.ts
@@ -1,0 +1,14 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import type { EsqlEsqlColumnInfo } from '@elastic/elasticsearch/lib/api/types';
+
+/** Find a result column by lowercase name against an allowlist of aliases. */
+export const findColumnIndex = (
+  columns: EsqlEsqlColumnInfo[],
+  aliases: readonly string[]
+): number => columns.findIndex((column) => aliases.includes(column.name.toLowerCase()));
